Add files and archives to the DistributedCache via configuraiton.
This patch enables you to add files and archives to the Distributed Cache via commandline specification and configuration within a job. This is important when mappers and reducers need to refer to configuration or relatively small data sets as part of their execution.

The new file "wordcount7.clj" demonstrates both of these two methods, and provides unit testing. The commandline usage can be tested manually, and the current additional test in the file tests specifying the added test files to the cache for use in the mapper.

The commandline help has been updated, as well as the README.txt with regards to wordcount7.clj.